[DV-172] refactor: 프로필 이미지 처리 수정
