+++ src/main/java/org/richardstallman/dvback/domain/user/controller/UserController.java
@@ -3,8 +3,6 @@ package org.richardstallman.dvback.domain.user.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.richardstallman.dvback.common.DvApiResponse;
@@ -12,7 +10,6 @@ import org.richardstallman.dvback.domain.user.converter.UserConverter;
 import org.richardstallman.dvback.domain.user.domain.request.UserRequestDto;
 import org.richardstallman.dvback.domain.user.domain.response.UserLoginResponseDto;
 import org.richardstallman.dvback.domain.user.domain.response.UserResponseDto;
-import org.richardstallman.dvback.domain.user.service.UserPostService;
 import org.richardstallman.dvback.domain.user.service.UserService;
 import org.richardstallman.dvback.global.jwt.JwtUtil;
 import org.richardstallman.dvback.global.oauth.service.TokenService;
@@ -20,7 +17,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @RestController
@@ -30,7 +26,6 @@ public class UserController {
 
   private final TokenService tokenService;
   private final UserService userService;
-  private final UserPostService userPostService;
   private final UserConverter userConverter;
   private final JwtUtil jwtUtil;
 
@@ -43,14 +38,11 @@ public class UserController {
     return ResponseEntity.ok("Logged out successfully");
   }
 
-  @PostMapping(value = "/info", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+  @PostMapping(value = "/info")
   public ResponseEntity<DvApiResponse<UserResponseDto>> updateUserInfo(
       @AuthenticationPrincipal Long userId,
-      @RequestPart("userInfo") @Valid final UserRequestDto userRequestDto,
-      @RequestPart("profileImage") @NotNull MultipartFile profileImage)
-      throws IOException {
-    final UserResponseDto userResponseDto =
-        userPostService.updateUserInfo(userId, userRequestDto, profileImage);
+      @RequestBody @Valid final UserRequestDto userRequestDto) {
+    final UserResponseDto userResponseDto = userService.updateUserInfo(userId, userRequestDto);
     return ResponseEntity.ok(DvApiResponse.of(userResponseDto));
   }
