+++ src/main/java/org/richardstallman/dvback/domain/user/domain/request/UserRequestDto.java
@@ -5,6 +5,7 @@ import java.time.LocalDate;
 import org.richardstallman.dvback.common.constant.CommonConstants;
 
 public record UserRequestDto(
+    @NotNull(message = "S3ProfileImageUrl is required") String s3ProfileImageUrl,
     @NotNull(message = "Name is required") String name,
     @NotNull(message = "Username is required") String username,
     @NotNull(message = "Nickname is required") String nickname,
