+++ src/main/java/org/richardstallman/dvback/domain/user/service/UserService.java
@@ -1,6 +1,7 @@
 package org.richardstallman.dvback.domain.user.service;
 
 import jakarta.servlet.http.HttpServletResponse;
+import org.richardstallman.dvback.domain.user.domain.request.UserRequestDto;
 import org.richardstallman.dvback.domain.user.domain.response.UserResponseDto;
 
 public interface UserService {
@@ -11,5 +12,5 @@ public interface UserService {
 
   boolean existsByUsername(String username);
 
-  String getProfileImage(Long userId);
+  UserResponseDto updateUserInfo(Long userId, UserRequestDto userRequestDto);
 }
